[Repository] Added AdvancedPageFilter
AdvancedPageFilter is an example of a BasicPageFilter implementation
overwriting the method toFilter().

This filter is based on the BasicPageFilter, but the value splits by
whitespaces and the result has to match all the value splits in any of
his columns (fields). You can add quotes "" to process any words like
only one.